fix: revert types
