+++ src/types.ts
@@ -1,7 +1,79 @@
-import type {NextFunction, Request, Response, Router} from 'express';
+import type {AppContext} from '@gravity-ui/nodekit';
+import type bodyParser from 'body-parser';
+import type {
+    ErrorRequestHandler,
+    NextFunction,
+    Request,
+    RequestHandler,
+    Response,
+    Router,
+} from 'express';
 
 import type {CSPPreset} from './csp';
 import type {getDefaultPresets} from './csp/default-presets';
+import type {CSPMiddlewareParams} from './csp/middleware';
+
+declare global {
+    // eslint-disable-next-line
+    namespace Express {
+        export interface Request {
+            id: string;
+            ctx: AppContext;
+            originalContext: AppContext;
+            routeInfo: AppRouteParams;
+        }
+    }
+}
+
+declare module 'express' {
+    export interface Request {
+        id: string;
+        ctx: AppContext;
+        originalContext: AppContext;
+        routeInfo: AppRouteParams;
+    }
+}
+
+declare module '@gravity-ui/nodekit' {
+    interface AppConfig {
+        expressTrustProxyNumber?: number | boolean;
+        expressCookieSecret?: string | string[];
+        expressRequestIdHeaderName?: string;
+
+        expressDisableBodyParserJSON?: boolean;
+        expressBodyParserJSONConfig?: bodyParser.OptionsJson;
+        expressExtendedBodyParserJSONLimitError?: boolean;
+
+        expressDisableBodyParserURLEncoded?: boolean;
+        expressBodyParserURLEncodedConfig?: bodyParser.OptionsUrlencoded;
+
+        expressBodyParserRawConfig?: bodyParser.Options;
+
+        appPort?: number;
+        appSocket?: string;
+
+        appFinalErrorHandler?: ErrorRequestHandler;
+        appAuthHandler?: RequestHandler;
+        appAuthPolicy?: `${AuthPolicy}`;
+
+        appBeforeAuthMiddleware?: RequestHandler[];
+        appAfterAuthMiddleware?: RequestHandler[];
+
+        appTelemetryChEnableSelfStats?: boolean;
+
+        expressCspEnable?: boolean;
+        expressCspPresets?:
+            | CSPPreset
+            | ((params: {getDefaultPresets: typeof getDefaultPresets}) => CSPPreset);
+        expressCspReportOnly?: boolean;
+        expressCspReportTo?: CSPMiddlewareParams['reportTo'];
+        expressCspReportUri?: CSPMiddlewareParams['reportUri'];
+
+        allowedLangs?: string[];
+        defaultLang?: string;
+        getLangByHostname?: (hostname: string) => string | undefined;
+    }
+}
 
 export enum AuthPolicy {
     disabled = 'disabled',
